hiver_openapi::macros

Module openapi

Rust implementation of Openapi Spec V3.1.

Modules

content

Implements content object for request body and response.

encoding

Implements encoding object for content.

example

Implements OpenAPI Example Object can be used to define examples for Responses and RequestBodys.

extensions

Implements OpenAPI Extensions.

external_docs

Implements OpenAPI External Docs Object types.

header

Implements OpenAPI Header Object types.

info

Implements OpenAPI Metadata types.

link

Implements Open API Link Object for responses.

path

Implements OpenAPI Path Object types.

request_body

Implements OpenAPI Request Body types.

response

Implements OpenApi Responses.

schema

Implements OpenAPI Schema Object types which can be used to define field properties, enum values, array or object types.

security

Implements OpenAPI Security Schema types.

server

Implements OpenAPI Server Object types to configure target servers.

tag

Implements OpenAPI Tag Object types.

xml

Implements OpenAPI Xml Object types.

Structs

AllOf

AllOf Composite Object component holds multiple components together where API endpoint will return a combination of all of them.

AllOfBuilder

Builder for AllOf with chainable configuration methods to create a new AllOf.

Array

Array represents Vec or slice type of items.

ArrayBuilder

Builder for Array with chainable configuration methods to create a new Array.

Components

Implements OpenAPI Components Object which holds supported reusable objects.

ComponentsBuilder

Builder for Components with chainable configuration methods to create a new Components.

Contact

OpenAPI Contact information of the API.

ContactBuilder

Builder for Contact with chainable configuration methods to create a new Contact. See the InfoBuilder for combined usage example.

Content

Content holds request body content or response content.

ContentBuilder

Builder for Content with chainable configuration methods to create a new Content.

Discriminator

OpenAPI Discriminator object which can be optionally used together with OneOf composite object.

ExternalDocs

Reference of external resource allowing extended documentation.

Header

Implements OpenAPI Header Object for response headers.

HeaderBuilder

Builder for Header with chainable configuration methods to create a new Header.

Info

OpenAPI Info object represents metadata of the API.

InfoBuilder

Builder for Info with chainable configuration methods to create a new Info.

License

OpenAPI License information of the API.

LicenseBuilder

Builder for License with chainable configuration methods to create a new License.

Object

Implements subset of OpenAPI Schema Object which allows adding other Schemas as properties to this Schema.

ObjectBuilder

Builder for Object with chainable configuration methods to create a new Object.

OneOf

OneOf Composite Object component holds multiple components together where API endpoint could return any of them.

OneOfBuilder

Builder for OneOf with chainable configuration methods to create a new OneOf.

OpenApi

Root object of the OpenAPI document.

OpenApiBuilder

Builder for OpenApi with chainable configuration methods to create a new OpenApi.

PathItem

Implements OpenAPI Path Item Object what describes Operations available on a single path.

Paths

Implements OpenAPI Paths Object.

PathsBuilder

Builder for Paths with chainable configuration methods to create a new Paths.

Ref

Implements OpenAPI Reference Object that can be used to reference reusable components such as Schemas or Responses.

Response

Implements OpenAPI Response Object.

ResponseBuilder

Builder for Response with chainable configuration methods to create a new Response.

Responses

Implements OpenAPI Responses Object.

ResponsesBuilder

Builder for Responses with chainable configuration methods to create a new Responses.

SecurityRequirement

OpenAPI security requirement object.

Server

Represents target server object. It can be used to alter server connection for path operations.

ServerBuilder

Builder for Server with chainable configuration methods to create a new Server.

ServerVariable

Implements OpenAPI Server Variable used to substitute variables in Server::url.

ServerVariableBuilder

Builder for ServerVariable with chainable configuration methods to create a new ServerVariable.

Tag

Implements OpenAPI Tag Object.

Enums

Deprecated

Value used to indicate whether reusable schema, parameter or operation is deprecated.

HttpMethod

HTTP method of the operation.

KnownFormat

Known schema format modifier property to provide fine detail of the primitive type.

OpenApiVersion

Represents available OpenAPI versions.

RefOr

A Ref or some other type T.

Required

Value used to indicate whether parameter or property is required.

Schema

Is super type for OpenAPI Schema Object. Schema is reusable resource what can be referenced from path operations and other components using Ref.

SchemaFormat

Additional format for SchemaType to fine tune the data type used. If the format is not supported by the UI it may default back to SchemaType alone. Format is an open value, so you can use any formats, even not those defined by the OpenAPI Specification.

Type

Represents data type fragment of Schema.

Traits

ToArray

This convenience trait allows quick way to wrap any RefOr<Schema> with Array schema.