hiver_openapi::macros::openapi::info

Struct InfoBuilder

pub struct InfoBuilder { /* private fields */ }

Builder for Info with chainable configuration methods to create a new Info.

Examples

Create Info using InfoBuilder.

let info = InfoBuilder::new()
     .title("My api")
     .version("1.0.0")
     .contact(Some(ContactBuilder::new()
          .name(Some("Admin Admin"))
          .email(Some("amdin@petapi.com"))
          .build()
      ))
     .build();

Implementations

impl InfoBuilder

pub fn new() -> InfoBuilder

Constructs a new InfoBuilder.

pub fn build(self) -> Info

Constructs a new Info taking all fields values from this object.

impl InfoBuilder

pub fn title<I>(self, title: I) -> InfoBuilder
where I: Into<String>,

Add title of the API.

pub fn version<I>(self, version: I) -> InfoBuilder
where I: Into<String>,

Add version of the api document typically the API version.

pub fn description<S>(self, description: Option<S>) -> InfoBuilder
where S: Into<String>,

Add description of the API.

pub fn terms_of_service<S>(self, terms_of_service: Option<S>) -> InfoBuilder
where S: Into<String>,

Add url for terms of the API.

pub fn contact(self, contact: Option<Contact>) -> InfoBuilder

Add contact information of the API.

pub fn license(self, license: Option<License>) -> InfoBuilder

Add license of the API.

pub fn extensions(self, extensions: Option<Extensions>) -> InfoBuilder

Add openapi extensions (x-something) of the API.