hiver_openapi::macros::openapi

Struct ObjectBuilder

pub struct ObjectBuilder { /* private fields */ }

Builder for Object with chainable configuration methods to create a new Object.

Implementations

impl ObjectBuilder

pub fn new() -> ObjectBuilder

Constructs a new ObjectBuilder.

pub fn build(self) -> Object

Constructs a new Object taking all fields values from this object.

impl ObjectBuilder

pub fn schema_type<T>(self, schema_type: T) -> ObjectBuilder
where T: Into<SchemaType>,

Add or change type of the object e.g. to change type to string use value SchemaType::Type(Type::String).

pub fn format(self, format: Option<SchemaFormat>) -> ObjectBuilder

Add or change additional format for detailing the schema type.

pub fn property<S, I>(self, property_name: S, component: I) -> ObjectBuilder
where S: Into<String>, I: Into<RefOr<Schema>>,

Add new property to the Object.

Method accepts property name and property component as an arguments.

pub fn additional_properties<I>( self, additional_properties: Option<I>, ) -> ObjectBuilder
where I: Into<AdditionalProperties<Schema>>,

Add additional Schema for non specified fields (Useful for typed maps).

pub fn property_names<S>(self, property_name: Option<S>) -> ObjectBuilder
where S: Into<Schema>,

Add additional Schema to describe property names of an object such as a map. See more details https://json-schema.org/draft/2020-12/draft-bhutton-json-schema-01#name-propertynames

pub fn required<I>(self, required_field: I) -> ObjectBuilder
where I: Into<String>,

Add field to the required fields of Object.

pub fn title<I>(self, title: Option<I>) -> ObjectBuilder
where I: Into<String>,

Add or change the title of the Object.

pub fn description<I>(self, description: Option<I>) -> ObjectBuilder
where I: Into<String>,

Add or change description of the property. Markdown syntax is supported.

pub fn default(self, default: Option<Value>) -> ObjectBuilder

Add or change default value for the object which is provided when user has not provided the input in Swagger UI.

pub fn deprecated(self, deprecated: Option<Deprecated>) -> ObjectBuilder

Add or change deprecated status for Object.

pub fn enum_values<I, E>(self, enum_values: Option<I>) -> ObjectBuilder
where I: IntoIterator<Item = E>, E: Into<Value>,

Add or change enum property variants.

pub fn example(self, example: Option<Value>) -> ObjectBuilder

👎Deprecated:
Since OpenAPI 3.1 prefer using examples

Add or change example shown in UI of the value for richer documentation.

Deprecated since 3.0.x. Prefer Object::examples instead

pub fn examples<I, V>(self, examples: I) -> ObjectBuilder
where I: IntoIterator<Item = V>, V: Into<Value>,

Add or change examples shown in UI of the value for richer documentation.

pub fn write_only(self, write_only: Option<bool>) -> ObjectBuilder

Add or change write only flag for Object.

pub fn read_only(self, read_only: Option<bool>) -> ObjectBuilder

Add or change read only flag for Object.

pub fn xml(self, xml: Option<Xml>) -> ObjectBuilder

Add or change additional Xml formatting of the Object.

pub fn multiple_of<N>(self, multiple_of: Option<N>) -> ObjectBuilder
where N: Into<Number>,

Set or change multiple_of validation flag for number and integer type values.

pub fn maximum<N>(self, maximum: Option<N>) -> ObjectBuilder
where N: Into<Number>,

Set or change inclusive maximum value for number and integer values.

pub fn minimum<N>(self, minimum: Option<N>) -> ObjectBuilder
where N: Into<Number>,

Set or change inclusive minimum value for number and integer values.

pub fn exclusive_maximum<N>(self, exclusive_maximum: Option<N>) -> ObjectBuilder
where N: Into<Number>,

Set or change exclusive maximum value for number and integer values.

pub fn exclusive_minimum<N>(self, exclusive_minimum: Option<N>) -> ObjectBuilder
where N: Into<Number>,

Set or change exclusive minimum value for number and integer values.

pub fn max_length(self, max_length: Option<usize>) -> ObjectBuilder

Set or change maximum length for string values.

pub fn min_length(self, min_length: Option<usize>) -> ObjectBuilder

Set or change minimum length for string values.

pub fn pattern<I>(self, pattern: Option<I>) -> ObjectBuilder
where I: Into<String>,

Set or change a valid regular expression for string value to match.

pub fn max_properties(self, max_properties: Option<usize>) -> ObjectBuilder

Set or change maximum number of properties the Object can hold.

pub fn min_properties(self, min_properties: Option<usize>) -> ObjectBuilder

Set or change minimum number of properties the Object can hold.

pub fn extensions(self, extensions: Option<Extensions>) -> ObjectBuilder

Add openapi extensions (x-something) for Object.

pub fn content_encoding<S>(self, content_encoding: S) -> ObjectBuilder
where S: Into<String>,

Set of change Object::content_encoding. Typically left empty but could be base64 for example.

pub fn content_media_type<S>(self, content_media_type: S) -> ObjectBuilder
where S: Into<String>,

Set of change Object::content_media_type. Value must be valid MIME type e.g. application/json.

pub fn to_array_builder(self) -> ArrayBuilder

Construct a new ArrayBuilder with this component set to ArrayBuilder::items.