hiver_openapi::macros::openapi

Struct OneOfBuilder

pub struct OneOfBuilder { /* private fields */ }

Builder for OneOf with chainable configuration methods to create a new OneOf.

Implementations

impl OneOfBuilder

pub fn new() -> OneOfBuilder

Constructs a new OneOfBuilder.

pub fn build(self) -> OneOf

Constructs a new OneOf taking all fields values from this object.

impl OneOfBuilder

pub fn item<I>(self, component: I) -> OneOfBuilder
where I: Into<RefOr<Schema>>,

Adds a given Schema to OneOf Composite Object.

pub fn schema_type<T>(self, schema_type: T) -> OneOfBuilder
where T: Into<SchemaType>,

Add or change type of the object e.g. to change type to string use value SchemaType::Type(Type::String).

pub fn title<I>(self, title: Option<I>) -> OneOfBuilder
where I: Into<String>,

Add or change the title of the OneOf.

pub fn description<I>(self, description: Option<I>) -> OneOfBuilder
where I: Into<String>,

Add or change optional description for OneOf component.

pub fn default(self, default: Option<Value>) -> OneOfBuilder

Add or change default value for the object which is provided when user has not provided the input in Swagger UI.

pub fn example(self, example: Option<Value>) -> OneOfBuilder

👎Deprecated:
Since OpenAPI 3.1 prefer using examples

Add or change example shown in UI of the value for richer documentation.

Deprecated since 3.0.x. Prefer OneOfBuilder::examples instead

pub fn examples<I, V>(self, examples: I) -> OneOfBuilder
where I: IntoIterator<Item = V>, V: Into<Value>,

Add or change examples shown in UI of the value for richer documentation.

pub fn discriminator(self, discriminator: Option<Discriminator>) -> OneOfBuilder

Add or change discriminator field of the composite OneOf type.

pub fn extensions(self, extensions: Option<Extensions>) -> OneOfBuilder

Add openapi extensions (x-something) for OneOf.

pub fn to_array_builder(self) -> ArrayBuilder

Construct a new ArrayBuilder with this component set to ArrayBuilder::items.