hiver_openapi::macros

Trait ToResponse

pub trait ToResponse<'__r> {
    // Required method
    fn response() -> (&'__r str, RefOr<Response>);
}

This trait is implemented to document a type which represents a single response which can be referenced or reused as a component in multiple operations.

ToResponse trait can also be derived with #[derive(ToResponse)].

Examples

use utoipa::{
    openapi::{RefOr, Response, ResponseBuilder},
    ToResponse,
};

struct MyResponse;

impl<'__r> ToResponse<'__r> for MyResponse {
    fn response() -> (&'__r str, RefOr<Response>) {
        (
            "MyResponse",
            ResponseBuilder::new().description("My Response").build().into(),
        )
    }
}

Required Methods

fn response() -> (&'__r str, RefOr<Response>)

Returns a tuple of response component name (to be referenced) to a response.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.